add header file, start working on folder conversion
