+++ src/main.cpp
@@ -6,98 +6,7 @@
 //  Copyright © 2019 Marco Janker. All rights reserved.
 //
 
-#define VERSION 4
-
-//Debug
-#define USEDEBUG 0
-#ifdef USEDEBUG
-#define Debug(x) std::cout << x << std::endl
-#else
-#define Debug(x)
-#endif
-
-
-#ifdef _WIN32
-#include <windows.h>
-#else
-#include <sys/ioctl.h>
-#endif
-
-#define cimg_display 0
-#define cimg_use_jpeg
-#define cimg_use_png
-#include "CImg.h"
-using namespace cimg_library;
-
-#include <cstring>
-#include <cstdio>
-#include <cstdint>
-#include <iostream>
-#include <fstream>
-#include <cstdint>
-#include <vector>
-#include <array>
-#include <thread>
-#include <future>
-#include <chrono>
-#include <algorithm> 
-
-// Boost
-#include <boost/regex.hpp>
-#include <boost/filesystem.hpp>
-#include <boost/filesystem/fstream.hpp>
-//#include <boost/timer/timer.hpp>
-using namespace boost::filesystem;
-//namespace po = boost::program_options;
-
-
-#include "ProgressBar.hpp"
-using namespace std;
-
-
-// Declare colors
-const unsigned char
-	white[]  = { 255, 255, 255 },
-	black[] = { 0, 0, 0 },
-	red[] = { 255, 0, 0 },
-	yellow[] = { 200, 155, 0 },
-	green[] = { 0, 255, 0 },
-	purple[] = { 175, 32, 186 },
-	blue[] = { 0, 0, 255 },
-	grey[] = { 127, 127, 127 };
-
-enum IMAGE_TYPE {
-	GIF = 0, PNG, BMP, JPG, JPEG, TIFF, TGA,
-	LAST // used for looping
-};
-
-// Declare variables
-int newimage_width = 800;
-int newimage_height = 600;
-const float coveragePercent = 2.0;
-const float tolerance = 0.1;
-const int backgroundColor = 255;
-const int posX = 0;
-const int posY = 0;
-const char * orderNumber;
-int textboxWidth;
-int textboxHeight;
-int textboxSize;
-int resizedImageSize;
-int textSize = 0;
-string strOrderNumber;
-int width;
-// Define number of steps to be reported in progress
-int steps = 5;
-
-char resizedAppendix[] = "_resized";
-char destDir[] = "Originale";
-string orderNumberFile = "overlay.txt";
-bool useOverlay = false;
-CImg<unsigned char> originalImage;
-CImg<unsigned char> resizedImage;
-CImg<unsigned char> textbackground;
-CImg<unsigned char> textbox;
+#include "resizer.hpp"
 
 CImg<unsigned char> resizeKeepAspectRatio(CImg<unsigned char> srcImage, const float dstWidth, const float dstHeight) {
 	float scaleHeight;
@@ -191,7 +100,7 @@ bool isValidImage (const std::string filename) {
     return false;
 }
 
-string getOrderNumber(path filePath) {
+string getOrderNumber(fs::path filePath) {
 	try {
 		//Check if a text document with name "overlay.txt" exists and use the containing number as order number
 		filePath = filePath / orderNumberFile;
@@ -254,6 +163,25 @@ float getTextSize (const char * text, int initialSize, float coveredSize, float 
     return size;
 }
 	
+std::vector<fs::path> getFolderContent(std::string path) {
+	
+	fs::path mypath(path);
+	fs::directory_iterator end_iter;
+
+	std::vector<fs::path> result_set;
+
+	if ( fs::exists(mypath) && fs::is_directory(mypath))
+	{
+		for( fs::directory_iterator dir_iter(mypath) ; dir_iter != end_iter ; ++dir_iter)
+		{
+			if (fs::is_regular_file(dir_iter->status()) )
+			{
+				result_set.push_back( *dir_iter );
+			}
+		}
+	}
+	return result_set;
+}
 int main(int argc, char * argv[]) {
 
 	std::cout <<  "Version: " << VERSION << std::endl;
@@ -261,6 +189,13 @@ int main(int argc, char * argv[]) {
 		std::cout << "missing filename" << std::endl;
 		exit(1);
 	}
+
+	// populate vector with content of folder
+	std::vector<fs::path> dirListing = getFolderContent(argc>1? argv[1] : ".");
+
+	for (std::vector<fs::path>::iterator it = dirListing.begin() ; it != dirListing.end(); ++it)
+    	std::cout << ' ' << *it;
+  	std::cout << '\n';
 
 	#ifdef _WIN32
 		CONSOLE_SCREEN_BUFFER_INFO csbi;
@@ -275,6 +210,11 @@ int main(int argc, char * argv[]) {
 	unsigned int total = steps * (argc - 1);
 	ProgressBar progressBar(total, width, '#', '-');
 	
+	// One argument could be a folder
+
+	// Otherwise it is a picture?
+
+	// Iterate over arguments
 	for (++argv; *argv; ++argv) {
 		
 		char* originalFile = *argv;
@@ -284,10 +224,13 @@ int main(int argc, char * argv[]) {
 		if ( ! isValidImage(strOriginalFile) ) { cout << "Not an image!" << endl; continue; }
 					
 		// Create path objects
-		path pathOriginalFile(originalFile);
-		path pathDestDirectory = pathOriginalFile.parent_path() / destDir;
-		path pathDestFile = pathOriginalFile.parent_path() / destDir / pathOriginalFile.filename();
-		path pathResizedFile = pathOriginalFile.parent_path() / (path(pathOriginalFile.stem().string() + resizedAppendix ).string() + pathOriginalFile.extension().string());
+		// path to original file
+		fs::path pathOriginalFile(originalFile);
+		// path object to backup copy of original image file
+		fs::path pathDestDirectory = pathOriginalFile.parent_path() / destDir;
+		fs::path pathDestFile = pathOriginalFile.parent_path() / destDir / pathOriginalFile.filename();
+		// path object to resized image file
+		fs::path pathResizedFile = pathOriginalFile.parent_path() / (fs::path(pathOriginalFile.stem().string() + resizedAppendix ).string() + pathOriginalFile.extension().string());
 		
 		// read order number once
 		if ( strOrderNumber.length() == 0 ) {
@@ -369,7 +312,7 @@ int main(int argc, char * argv[]) {
 							try {
 								boost::filesystem::create_directory(pathDestDirectory);
 							}
-							catch (const filesystem_error& ex)
+							catch (const fs::filesystem_error& ex)
 							{
 								std::fprintf(stderr,"Filesystem Error: %s",ex.what());
 								system("pause");
@@ -384,7 +327,7 @@ int main(int argc, char * argv[]) {
 						try {
 							boost::filesystem::rename(pathOriginalFile,pathDestFile);
 						}
-						catch (const filesystem_error& ex)
+						catch (const fs::filesystem_error& ex)
 						{
 							std::fprintf(stderr,"Filesystem Error: %s",ex.what());
 							system("pause");
